Always show the list of documents, even if there are 0 or 1 documents
It is a bit jarring to have the list appear and disappear, and the DocumentPage shift/adjust in size, when the second page is scanned or deleted. It seems like it wouldn't be much of a functionality regression to just always show the list, since it's just reducing the potential size of the image on the screen by a bit, and there are still zoom options available.

But, in case someone wants to hide the list view, it is now adjustable in size down to 0px. The handle is also increased in size a bit to make it easier to grab.