Ask before overwriting in ExportDialog
If the user manually changes the filename in the TextField and clicks
"Save", the program will overwrite existing files without confirmation.

Solution: Remove the file path from the fields in the ExportDialog and
move the possibility to change the path in the FileDialog. This way the
check to avoid overwriting files is done in a single place.

BUG: 508559
ENHANCEMENT: 490999
FIXED-IN: 26.04.0